requirements.txt
csv
mysql-connector-python